Update `.github/`, `cicd/`, `pkg/` and `tools/` to latest 3006.x
