+++ .github/workflows/build-salt-onedir.yml
@@ -23,21 +23,11 @@ on:
         type: string
         default: 0.12.3
         description: The version of relenv to use
-      python-version-linux:
+      python-version:
         required: false
         type: string
         default: 3.10.9
         description: The version of python to use with relenv
-      python-version-macos:
-        required: false
-        type: string
-        default: 3.10.9
-        description: The version of python to use with relenv
-      python-version-windows:
-        required: false
-        type: string
-        description: The version of python to use with relenv on Windows
-        default: 3.10.9
 
 env:
   RELENV_DATA: "${{ github.workspace }}/.relenv"
@@ -77,7 +67,7 @@ jobs:
           arch: ${{ matrix.arch }}
           version: ${{ inputs.relenv-version }}
           cache-seed: ${{ inputs.cache-seed }}
-          python-version: ${{ inputs.python-version-linux }}
+          python-version: ${{ inputs.python-version }}
 
       - name: Install Salt into Relenv Onedir
         uses: ./.github/actions/build-onedir-salt
@@ -85,7 +75,7 @@ jobs:
           platform: linux
           arch: ${{ matrix.arch }}
           salt-version: "${{ inputs.salt-version }}"
-          python-version: "${{ inputs.python-version-linux }}"
+          python-version: "${{ inputs.python-version }}"
           cache-prefix: ${{ inputs.cache-seed }}|relenv|${{ steps.setup-relenv.outputs.version }}
 
 
@@ -115,7 +105,7 @@ jobs:
           arch: ${{ matrix.arch }}
           version: ${{ inputs.relenv-version }}
           cache-seed: ${{ inputs.cache-seed }}
-          python-version: ${{ inputs.python-version-windows }}
+          python-version: ${{ inputs.python-version }}
 
       - name: Setup Python Tools Scripts
         uses: ./.github/actions/setup-python-tools-scripts
@@ -132,7 +122,7 @@ jobs:
           platform: windows
           arch: ${{ matrix.arch }}
           salt-version: "${{ inputs.salt-version }}"
-          python-version: "${{ inputs.python-version-windows }}"
+          python-version: "${{ inputs.python-version }}"
           cache-prefix: ${{ inputs.cache-seed }}|relenv|${{ steps.setup-relenv.outputs.version }}
 
 
@@ -161,7 +151,7 @@ jobs:
           arch: ${{ matrix.arch }}
           version: ${{ inputs.relenv-version }}
           cache-seed: ${{ inputs.cache-seed }}
-          python-version: ${{ inputs.python-version-macos }}
+          python-version: ${{ inputs.python-version }}
 
       - name: Setup Python Tools Scripts
         uses: ./.github/actions/setup-python-tools-scripts
@@ -178,5 +168,5 @@ jobs:
           platform: darwin
           arch: ${{ matrix.arch }}
           salt-version: "${{ inputs.salt-version }}"
-          python-version: "${{ inputs.python-version-macos }}"
+          python-version: "${{ inputs.python-version }}"
           cache-prefix: ${{ inputs.cache-seed }}|relenv|${{ steps.setup-relenv.outputs.version }}
