+++ .github/workflows/templates/ci.yml.jinja
@@ -1,12 +1,6 @@
 <%- extends 'layout.yml.jinja' %>
 <%- set pre_commit_version = "3.0.4" %>
 
-<%- block on %>
-on:
-  push: {}
-  pull_request: {}
-<%- endblock on %>
-
 
 <%- block jobs %>
   <{- super() }>
@@ -58,8 +52,7 @@ on:
       - x86_64
     <%- else %>
     if: ${{ fromJSON(needs.prepare-workflow.outputs.jobs)['<{ job_name }>'] && fromJSON(needs.prepare-workflow.outputs.runners)['github-hosted'] }}
-    runs-on:
-      - ubuntu-latest
+    runs-on: ${{ github.event.repository.private && fromJSON('["self-hosted", "linux", "x86_64"]') || 'ubuntu-latest' }}
     <%- endif %>
     needs:
       - prepare-workflow
@@ -220,7 +213,7 @@ on:
     needs:
       - prepare-workflow
       - prepare-release
-    runs-on: ubuntu-latest
+    runs-on: ${{ github.event.repository.private && fromJSON('["self-hosted", "linux", "x86_64"]') || 'ubuntu-latest' }}
     steps:
       - uses: actions/checkout@v3
 
@@ -275,9 +268,7 @@ on:
       self-hosted-runners: ${{ fromJSON(needs.prepare-workflow.outputs.runners)['self-hosted'] }}
       github-hosted-runners: ${{ fromJSON(needs.prepare-workflow.outputs.runners)['github-hosted'] }}
       relenv-version: "<{ relenv_version }>"
-      python-version-linux: "<{ python_version_linux }>"
-      python-version-macos: "<{ python_version_macos }>"
-      python-version-windows: "<{ python_version_windows }>"
+      python-version: "<{ python_version }>"
 
   <%- endif %>
 
@@ -300,9 +291,7 @@ on:
       self-hosted-runners: ${{ fromJSON(needs.prepare-workflow.outputs.runners)['self-hosted'] }}
       github-hosted-runners: ${{ fromJSON(needs.prepare-workflow.outputs.runners)['github-hosted'] }}
       relenv-version: "<{ relenv_version }>"
-      python-version-linux: "<{ python_version_linux }>"
-      python-version-macos: "<{ python_version_macos }>"
-      python-version-windows: "<{ python_version_windows }>"
+      python-version: "<{ python_version }>"
 
   <%- endif %>
