+++ .github/workflows/templates/test-salt-pkg.yml.jinja
@@ -9,7 +9,6 @@
                                ("debian-10", "Debian 10", "x86_64", "deb"),
                                ("debian-11", "Debian 11", "x86_64", "deb"),
                                ("debian-11-arm64", "Debian 11 Arm64", "aarch64", "deb"),
-                               ("fedora-36", "Fedora 36", "x86_64", "rpm"),
                                ("fedora-37", "Fedora 37", "x86_64", "rpm"),
                                ("fedora-38", "Fedora 38", "x86_64", "rpm"),
                                ("ubuntu-20.04", "Ubuntu 20.04", "x86_64", "deb"),
@@ -51,9 +50,10 @@
       arch: <{ arch }>
       salt-version: "${{ needs.prepare-workflow.outputs.salt-version }}"
       pkg-type: <{ pkg_type }>
-      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version_linux }>
+      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version }>
       skip-code-coverage: <{ skip_test_coverage_check }>
       skip-junit-reports: <{ skip_junit_reports_check }>
+      testing-releases: ${{ needs.prepare-workflow.outputs.testing-releases }}
 
   <%- endfor %>
 
@@ -76,9 +76,10 @@
       arch: <{ arch }>
       salt-version: "${{ needs.prepare-workflow.outputs.salt-version }}"
       pkg-type: macos
-      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version_macos }>
+      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version }>
       skip-code-coverage: <{ skip_test_coverage_check }>
       skip-junit-reports: <{ skip_junit_reports_check }>
+      testing-releases: ${{ needs.prepare-workflow.outputs.testing-releases }}
 
   <%- endfor %>
 
@@ -103,9 +104,10 @@
       arch: <{ arch }>
       salt-version: "${{ needs.prepare-workflow.outputs.salt-version }}"
       pkg-type: <{ pkg_type }>
-      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version_windows }>
+      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version }>
       skip-code-coverage: <{ skip_test_coverage_check }>
       skip-junit-reports: <{ skip_junit_reports_check }>
+      testing-releases: ${{ needs.prepare-workflow.outputs.testing-releases }}
 
     <%- endfor %>
   <%- endfor %>
