+++ .github/workflows/templates/test-salt.yml.jinja
@@ -18,8 +18,7 @@
       arch: amd64
       testrun: ${{ needs.prepare-workflow.outputs.testrun }}
       salt-version: "${{ needs.prepare-workflow.outputs.salt-version }}"
-      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version_windows }>
-      pull-labels: ${{ needs.prepare-workflow.outputs.pull-labels }}
+      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version }>
       skip-code-coverage: <{ skip_test_coverage_check }>
       skip-junit-reports: <{ skip_junit_reports_check }>
 
@@ -42,8 +41,7 @@
       arch: x86_64
       testrun: ${{ needs.prepare-workflow.outputs.testrun }}
       salt-version: "${{ needs.prepare-workflow.outputs.salt-version }}"
-      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version_macos }>
-      pull-labels: ${{ needs.prepare-workflow.outputs.pull-labels }}
+      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version }>
       skip-code-coverage: <{ skip_test_coverage_check }>
       skip-junit-reports: <{ skip_junit_reports_check }>
 
@@ -59,7 +57,6 @@
                                        ("debian-10", "Debian 10", "x86_64"),
                                        ("debian-11", "Debian 11", "x86_64"),
                                        ("debian-11-arm64", "Debian 11 Arm64", "aarch64"),
-                                       ("fedora-36", "Fedora 36", "x86_64"),
                                        ("fedora-37", "Fedora 37", "x86_64"),
                                        ("fedora-38", "Fedora 38", "x86_64"),
                                        ("opensuse-15", "Opensuse 15", "x86_64"),
@@ -85,8 +82,7 @@
       arch: <{ arch }>
       testrun: ${{ needs.prepare-workflow.outputs.testrun }}
       salt-version: "${{ needs.prepare-workflow.outputs.salt-version }}"
-      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version_linux }>
-      pull-labels: ${{ needs.prepare-workflow.outputs.pull-labels }}
+      cache-prefix: ${{ needs.prepare-workflow.outputs.cache-seed }}|<{ python_version }>
       skip-code-coverage: <{ skip_test_coverage_check }>
       skip-junit-reports: <{ skip_junit_reports_check }>
