+++ cicd/amis.yml
@@ -1,1 +1,1 @@
-centosstream-9-x86_64: ami-044545f7a74d46acc
+centosstream-9-x86_64: ami-0bd92f4dca5d74017
