+++ pkg/tests/conftest.py
@@ -44,6 +44,21 @@ def grains(sminion):
     return sminion.opts["grains"].copy()
 
 
+@pytest.fixture(scope="session", autouse=True)
+def _system_up_to_date(
+    grains,
+    shell,
+):
+    if grains["os_family"] == "Debian":
+        ret = shell.run("apt", "update")
+        assert ret.returncode == 0
+        ret = shell.run("apt", "upgrade", "-y")
+        assert ret.returncode == 0
+    elif grains["os_family"] == "Redhat":
+        ret = shell.run("yum", "update", "-y")
+        assert ret.returncode == 0
+
+
 def pytest_addoption(parser):
     """
     register argparse-style options and ini-style config values.
@@ -348,6 +363,8 @@ def salt_master(salt_factories, install_salt, state_tree, pillar_tree):
             master_script = True
         # this check will need to be changed to install_salt.relenv
         # once the package version returns 3006 and not 3005 on master
+        if install_salt.relenv:
+            master_script = True
         elif not install_salt.upgrade:
             master_script = True
 
@@ -360,6 +377,8 @@ def salt_master(salt_factories, install_salt, state_tree, pillar_tree):
         python_executable = install_salt.bin_dir / "Scripts" / "python.exe"
         if install_salt.classic:
             python_executable = install_salt.bin_dir / "python.exe"
+        if install_salt.relenv:
+            python_executable = install_salt.install_dir / "Scripts" / "python.exe"
         factory = salt_factories.salt_master_daemon(
             random_string("master-"),
             defaults=config_defaults,
