+++ pkg/tests/download/test_pkg_download.py
@@ -81,6 +81,11 @@ def root_url(salt_release):
     return _root_url
 
 
+@pytest.fixture(scope="module")
+def package_type():
+    return os.environ.get("DOWNLOAD_TEST_PACKAGE_TYPE")
+
+
 def get_salt_release():
     salt_release = os.environ.get("SALT_RELEASE")
     pkg_test_type = os.environ.get("PKG_TEST_TYPE", "install")
@@ -131,8 +136,23 @@ def salt_release():
 
 
 @pytest.fixture(scope="module")
+def onedir_install_path(tmp_path_factory):
+    install_path = tmp_path_factory.mktemp("onedir_install")
+    yield install_path
+    shutil.rmtree(install_path, ignore_errors=True)
+
+
+@pytest.fixture(scope="module")
 def _setup_system(
-    tmp_path_factory, grains, shell, root_url, salt_release, gpg_key_name, repo_subpath
+    grains,
+    shell,
+    root_url,
+    salt_release,
+    gpg_key_name,
+    repo_subpath,
+    package_type,
+    tmp_path_factory,
+    onedir_install_path,
 ):
     downloads_path = tmp_path_factory.mktemp("downloads")
     try:
@@ -144,6 +164,8 @@ def _setup_system(
                 salt_release=salt_release,
                 downloads_path=downloads_path,
                 repo_subpath=repo_subpath,
+                package_type=package_type,
+                onedir_install_path=onedir_install_path,
             ):
                 yield
         else:
@@ -154,6 +176,8 @@ def _setup_system(
                     salt_release=salt_release,
                     downloads_path=downloads_path,
                     repo_subpath=repo_subpath,
+                    package_type=package_type,
+                    onedir_install_path=onedir_install_path,
                 )
             elif grains["os"] == "Amazon":
                 setup_redhat_family(
@@ -210,6 +234,8 @@ def _setup_system(
                     downloads_path=downloads_path,
                     gpg_key_name=gpg_key_name,
                     repo_subpath=repo_subpath,
+                    package_type=package_type,
+                    onedir_install_path=onedir_install_path,
                 )
             else:
                 pytest.fail("Don't know how to handle %s", grains["osfinger"])
@@ -286,152 +312,237 @@ def setup_debian_family(
     downloads_path,
     gpg_key_name,
     repo_subpath,
+    package_type,
+    onedir_install_path,
 ):
     arch = os.environ.get("SALT_REPO_ARCH") or "amd64"
-    if arch == "aarch64":
-        arch = "arm64"
-    elif arch == "x86_64":
-        arch = "amd64"
-
     ret = shell.run("apt-get", "update", "-y", check=False)
     if ret.returncode != 0:
         pytest.fail(str(ret))
 
-    if repo_subpath == "minor":
-        repo_url_base = (
-            f"{root_url}/{os_name}/{os_version}/{arch}/{repo_subpath}/{salt_release}"
-        )
-    else:
-        repo_url_base = f"{root_url}/{os_name}/{os_version}/{arch}/{repo_subpath}"
-    gpg_file_url = f"{root_url}/{os_name}/{os_version}/{arch}/{gpg_key_name}"
+    if package_type == "package":
+        if arch == "aarch64":
+            arch = "arm64"
+        elif arch == "x86_64":
+            arch = "amd64"
 
-    try:
-        pytest.helpers.download_file(gpg_file_url, downloads_path / gpg_key_name)
-    except Exception as exc:
-        pytest.fail(f"Failed to download {gpg_file_url}: {exc}")
-
-    salt_sources_path = downloads_path / "salt.list"
-    salt_sources_path.write_text(
-        f"deb [signed-by=/usr/share/keyrings/{gpg_key_name} arch={arch}] {repo_url_base} {os_codename} main\n"
-    )
-    commands = [
-        (
-            "mv",
-            str(downloads_path / gpg_key_name),
-            f"/usr/share/keyrings/{gpg_key_name}",
-        ),
-        (
-            "mv",
-            str(salt_sources_path),
-            "/etc/apt/sources.list.d/salt.list",
-        ),
-        ("apt-get", "install", "-y", "ca-certificates"),
-        ("update-ca-certificates",),
-        ("apt-get", "update"),
-        (
-            "apt-get",
-            "install",
-            "-y",
-            "salt-master",
-            "salt-minion",
-            "salt-ssh",
-            "salt-syndic",
-            "salt-cloud",
-            "salt-api",
-        ),
-    ]
-    for cmd in commands:
-        ret = shell.run(*cmd)
-        if ret.returncode != 0:
-            pytest.fail(str(ret))
-
-
-def setup_macos(shell, root_url, salt_release, downloads_path, repo_subpath):
-    arch = os.environ.get("SALT_REPO_ARCH") or "x86_64"
-    if arch == "aarch64":
-        arch = "arm64"
-
-    if packaging.version.parse(salt_release) > packaging.version.parse("3005"):
-        mac_pkg = f"salt-{salt_release}-py3-{arch}.pkg"
         if repo_subpath == "minor":
-            mac_pkg_url = f"{root_url}/macos/{repo_subpath}/{salt_release}/{mac_pkg}"
+            repo_url_base = f"{root_url}/{os_name}/{os_version}/{arch}/{repo_subpath}/{salt_release}"
         else:
-            mac_pkg_url = f"{root_url}/macos/{repo_subpath}/{mac_pkg}"
+            repo_url_base = f"{root_url}/{os_name}/{os_version}/{arch}/{repo_subpath}"
+        gpg_file_url = f"{root_url}/{os_name}/{os_version}/{arch}/{gpg_key_name}"
+
+        try:
+            pytest.helpers.download_file(gpg_file_url, downloads_path / gpg_key_name)
+        except Exception as exc:
+            pytest.fail(f"Failed to download {gpg_file_url}: {exc}")
+
+        salt_sources_path = downloads_path / "salt.list"
+        salt_sources_path.write_text(
+            f"deb [signed-by=/usr/share/keyrings/{gpg_key_name} arch={arch}] {repo_url_base} {os_codename} main\n"
+        )
+        commands = [
+            (
+                "mv",
+                str(downloads_path / gpg_key_name),
+                f"/usr/share/keyrings/{gpg_key_name}",
+            ),
+            (
+                "mv",
+                str(salt_sources_path),
+                "/etc/apt/sources.list.d/salt.list",
+            ),
+            ("apt-get", "install", "-y", "ca-certificates"),
+            ("update-ca-certificates",),
+            ("apt-get", "update"),
+            (
+                "apt-get",
+                "install",
+                "-y",
+                "salt-master",
+                "salt-minion",
+                "salt-ssh",
+                "salt-syndic",
+                "salt-cloud",
+                "salt-api",
+            ),
+        ]
+        for cmd in commands:
+            ret = shell.run(*cmd)
+            if ret.returncode != 0:
+                pytest.fail(str(ret))
     else:
-        mac_pkg_url = f"{root_url}/macos/{salt_release}/{mac_pkg}"
-        mac_pkg = f"salt-{salt_release}-macos-{arch}.pkg"
+        # We are testing the onedir download
+        onedir_name = f"salt-{salt_release}-onedir-linux-{arch}.tar.xz"
+        if repo_subpath == "minor":
+            repo_url_base = f"{root_url}/onedir/{repo_subpath}/{salt_release}"
+        else:
+            repo_url_base = f"{root_url}/onedir/{repo_subpath}"
+        onedir_url = f"{repo_url_base}/{onedir_name}"
+        onedir_location = downloads_path / onedir_name
+        onedir_extracted = onedir_install_path
 
-    mac_pkg_path = downloads_path / mac_pkg
-    pytest.helpers.download_file(mac_pkg_url, mac_pkg_path)
+        try:
+            pytest.helpers.download_file(onedir_url, onedir_location)
+        except Exception as exc:
+            pytest.fail(f"Failed to download {onedir_url}: {exc}")
 
-    ret = shell.run(
-        "installer",
-        "-pkg",
-        str(mac_pkg_path),
-        "-target",
-        "/",
-        check=False,
-    )
-    assert ret.returncode == 0, ret
+        shell.run("tar", "xvf", str(onedir_location), "-C", str(onedir_extracted))
+
+
+def setup_macos(
+    shell,
+    root_url,
+    salt_release,
+    downloads_path,
+    repo_subpath,
+    package_type,
+    onedir_install_path,
+):
+    arch = os.environ.get("SALT_REPO_ARCH") or "x86_64"
+    if package_type == "package":
+        if arch == "aarch64":
+            arch = "arm64"
+
+        if packaging.version.parse(salt_release) > packaging.version.parse("3005"):
+            mac_pkg = f"salt-{salt_release}-py3-{arch}.pkg"
+            if repo_subpath == "minor":
+                mac_pkg_url = (
+                    f"{root_url}/macos/{repo_subpath}/{salt_release}/{mac_pkg}"
+                )
+            else:
+                mac_pkg_url = f"{root_url}/macos/{repo_subpath}/{mac_pkg}"
+        else:
+            mac_pkg_url = f"{root_url}/macos/{salt_release}/{mac_pkg}"
+            mac_pkg = f"salt-{salt_release}-macos-{arch}.pkg"
+
+        mac_pkg_path = downloads_path / mac_pkg
+        pytest.helpers.download_file(mac_pkg_url, mac_pkg_path)
+
+        ret = shell.run(
+            "installer",
+            "-pkg",
+            str(mac_pkg_path),
+            "-target",
+            "/",
+            check=False,
+        )
+        assert ret.returncode == 0, ret
+    else:
+        # We are testing the onedir download
+        onedir_name = f"salt-{salt_release}-onedir-darwin-{arch}.tar.xz"
+        if repo_subpath == "minor":
+            repo_url_base = f"{root_url}/onedir/{repo_subpath}/{salt_release}"
+        else:
+            repo_url_base = f"{root_url}/onedir/{repo_subpath}"
+        onedir_url = f"{repo_url_base}/{onedir_name}"
+        onedir_location = downloads_path / onedir_name
+        onedir_extracted = onedir_install_path
+
+        try:
+            pytest.helpers.download_file(onedir_url, onedir_location)
+        except Exception as exc:
+            pytest.fail(f"Failed to download {onedir_url}: {exc}")
+
+        shell.run("tar", "xvf", str(onedir_location), "-C", str(onedir_extracted))
 
 
 @contextlib.contextmanager
-def setup_windows(shell, root_url, salt_release, downloads_path, repo_subpath):
+def setup_windows(
+    shell,
+    root_url,
+    salt_release,
+    downloads_path,
+    repo_subpath,
+    package_type,
+    onedir_install_path,
+):
     try:
-        root_dir = pathlib.Path(r"C:\Program Files\Salt Project\Salt")
-
         arch = os.environ.get("SALT_REPO_ARCH") or "amd64"
-        install_type = os.environ.get("INSTALL_TYPE") or "msi"
-        if packaging.version.parse(salt_release) > packaging.version.parse("3005"):
-            if install_type.lower() == "nsis":
-                if arch.lower() != "x86":
-                    arch = arch.upper()
-                win_pkg = f"Salt-Minion-{salt_release}-Py3-{arch}-Setup.exe"
+        if package_type != "onedir":
+            root_dir = pathlib.Path(r"C:\Program Files\Salt Project\Salt")
+
+            if packaging.version.parse(salt_release) > packaging.version.parse("3005"):
+                if package_type.lower() == "nsis":
+                    if arch.lower() != "x86":
+                        arch = arch.upper()
+                    win_pkg = f"Salt-Minion-{salt_release}-Py3-{arch}-Setup.exe"
+                else:
+                    if arch.lower() != "x86":
+                        arch = arch.upper()
+                    win_pkg = f"Salt-Minion-{salt_release}-Py3-{arch}.msi"
+                if repo_subpath == "minor":
+                    win_pkg_url = (
+                        f"{root_url}/windows/{repo_subpath}/{salt_release}/{win_pkg}"
+                    )
+                else:
+                    win_pkg_url = f"{root_url}/windows/{repo_subpath}/{win_pkg}"
+                ssm_bin = root_dir / "ssm.exe"
             else:
-                if arch.lower() != "x86":
-                    arch = arch.upper()
-                win_pkg = f"Salt-Minion-{salt_release}-Py3-{arch}.msi"
-            if repo_subpath == "minor":
-                win_pkg_url = (
-                    f"{root_url}/windows/{repo_subpath}/{salt_release}/{win_pkg}"
+                win_pkg = f"salt-{salt_release}-windows-{arch}.exe"
+                win_pkg_url = f"{root_url}/windows/{salt_release}/{win_pkg}"
+                ssm_bin = root_dir / "bin" / "ssm_bin"
+
+            pkg_path = downloads_path / win_pkg
+
+            pytest.helpers.download_file(win_pkg_url, pkg_path)
+            if package_type.lower() == "nsis":
+                ret = shell.run(str(pkg_path), "/start-minion=0", "/S", check=False)
+            else:
+                ret = shell.run(
+                    "msiexec", "/qn", "/i", str(pkg_path), 'START_MINION=""'
                 )
+            assert ret.returncode == 0, ret
+
+            log.debug("Removing installed salt-minion service")
+            ret = shell.run(
+                "cmd",
+                "/c",
+                str(ssm_bin),
+                "remove",
+                "salt-minion",
+                "confirm",
+                check=False,
+            )
+            assert ret.returncode == 0, ret
+        else:
+            # We are testing the onedir download
+            onedir_name = f"salt-{salt_release}-onedir-windows-{arch}.zip"
+            if repo_subpath == "minor":
+                repo_url_base = f"{root_url}/onedir/{repo_subpath}/{salt_release}"
             else:
-                win_pkg_url = f"{root_url}/windows/{repo_subpath}/{win_pkg}"
-            ssm_bin = root_dir / "ssm.exe"
-        else:
-            win_pkg = f"salt-{salt_release}-windows-{arch}.exe"
-            win_pkg_url = f"{root_url}/windows/{salt_release}/{win_pkg}"
-            ssm_bin = root_dir / "bin" / "ssm_bin"
+                repo_url_base = f"{root_url}/onedir/{repo_subpath}"
+            onedir_url = f"{repo_url_base}/{onedir_name}"
+            onedir_location = downloads_path / onedir_name
+            onedir_extracted = onedir_install_path
 
-        pkg_path = downloads_path / win_pkg
+            try:
+                pytest.helpers.download_file(onedir_url, onedir_location)
+            except Exception as exc:
+                pytest.fail(f"Failed to download {onedir_url}: {exc}")
 
-        pytest.helpers.download_file(win_pkg_url, pkg_path)
-        if install_type.lower() == "nsis":
-            ret = shell.run(str(pkg_path), "/start-minion=0", "/S", check=False)
-        else:
-            ret = shell.run("msiexec", "/qn", "/i", str(pkg_path), 'START_MINION=""')
-        assert ret.returncode == 0, ret
-
-        log.debug("Removing installed salt-minion service")
-        ret = shell.run(
-            "cmd", "/c", str(ssm_bin), "remove", "salt-minion", "confirm", check=False
-        )
-        assert ret.returncode == 0, ret
+            shell.run("unzip", str(onedir_location), "-d", str(onedir_extracted))
         yield
     finally:
         # We need to uninstall the MSI packages, otherwise they will not install correctly
-        if install_type.lower() == "msi":
+        if package_type.lower() == "msi":
             ret = shell.run("msiexec", "/qn", "/x", str(pkg_path))
             assert ret.returncode == 0, ret
 
 
 @pytest.fixture(scope="module")
-def install_dir(_setup_system):
-    if platform.is_windows():
-        return pathlib.Path(os.getenv("ProgramFiles"), "Salt Project", "Salt").resolve()
-    if platform.is_darwin():
-        return pathlib.Path("/opt", "salt")
-    return pathlib.Path("/opt", "saltstack", "salt")
+def install_dir(_setup_system, package_type, onedir_install_path):
+    if package_type != "onedir":
+        if platform.is_windows():
+            return pathlib.Path(
+                os.getenv("ProgramFiles"), "Salt Project", "Salt"
+            ).resolve()
+        if platform.is_darwin():
+            return pathlib.Path("/opt", "salt")
+        return pathlib.Path("/opt", "saltstack", "salt")
+    else:
+        # We are testing the onedir
+        return onedir_install_path / "salt"
 
 
 @pytest.fixture(scope="module")
