+++ pkg/tests/integration/test_salt_user.py
@@ -1,5 +1,6 @@
 import pathlib
 import subprocess
+import sys
 
 import psutil
 import pytest
@@ -61,7 +62,9 @@ def test_salt_cloud_dirs(install_salt):
     Test the correct user is running the Salt Master
     """
     paths = [
-        "/opt/saltstack/salt/lib/python3.10/site-packages/salt/cloud/deploy",
+        "/opt/saltstack/salt/lib/python{}.{}/site-packages/salt/cloud/deploy".format(
+            *sys.version_info
+        ),
         "/etc/salt/cloud.deploy.d",
     ]
     for name in paths:
