+++ tools/ci.py
@@ -8,12 +8,19 @@ import json
 import logging
 import os
 import pathlib
+import random
+import sys
 import time
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 from ptscripts import Context, command_group
 
 import tools.utils
+
+if sys.version_info < (3, 11):
+    from typing_extensions import NotRequired, TypedDict
+else:
+    from typing import NotRequired, TypedDict  # pylint: disable=no-name-in-module
 
 log = logging.getLogger(__name__)
 
@@ -299,6 +306,23 @@ def define_jobs(
             )
         return
 
+    # This is a pull-request
+
+    labels: list[str] = []
+    gh_event_path = os.environ.get("GITHUB_EVENT_PATH") or None
+    if gh_event_path is not None:
+        try:
+            gh_event = json.loads(open(gh_event_path).read())
+        except Exception as exc:
+            ctx.error(
+                f"Could not load the GH Event payload from {gh_event_path!r}:\n", exc
+            )
+            ctx.exit(1)
+
+        labels.extend(
+            label[0] for label in _get_pr_test_labels_from_event_payload(gh_event)
+        )
+
     if not changed_files.exists():
         ctx.error(f"The '{changed_files}' file does not exist.")
         ctx.error(
@@ -348,9 +372,16 @@ def define_jobs(
         changed_files_contents["workflows"],
     }
     if jobs["test-pkg"] and required_pkg_test_changes == {"false"}:
-        with open(github_step_summary, "a", encoding="utf-8") as wfh:
-            wfh.write("De-selecting the 'test-pkg' job.\n")
-        jobs["test-pkg"] = False
+        if "test:pkg" in labels:
+            with open(github_step_summary, "a", encoding="utf-8") as wfh:
+                wfh.write(
+                    "The 'test-pkg' job is forcefully selected by the use of the 'test:pkg' label.\n"
+                )
+            jobs["test-pkg"] = True
+        else:
+            with open(github_step_summary, "a", encoding="utf-8") as wfh:
+                wfh.write("De-selecting the 'test-pkg' job.\n")
+            jobs["test-pkg"] = False
 
     if jobs["test-pkg-download"] and required_pkg_test_changes == {"false"}:
         with open(github_step_summary, "a", encoding="utf-8") as wfh:
@@ -379,6 +410,13 @@ def define_jobs(
     ctx.info("Writing 'jobs' to the github outputs file")
     with open(github_output, "a", encoding="utf-8") as wfh:
         wfh.write(f"jobs={json.dumps(jobs)}\n")
+
+
+class TestRun(TypedDict):
+    type: str
+    skip_code_coverage: bool
+    from_filenames: NotRequired[str]
+    selected_tests: NotRequired[dict[str, bool]]
 
 
 @ci.command(
@@ -415,10 +453,31 @@ def define_testrun(ctx: Context, event_name: str, changed_files: pathlib.Path):
     if TYPE_CHECKING:
         assert github_step_summary is not None
 
+    labels: list[str] = []
+    gh_event_path = os.environ.get("GITHUB_EVENT_PATH") or None
+    if gh_event_path is not None:
+        try:
+            gh_event = json.loads(open(gh_event_path).read())
+        except Exception as exc:
+            ctx.error(
+                f"Could not load the GH Event payload from {gh_event_path!r}:\n", exc
+            )
+            ctx.exit(1)
+
+        labels.extend(
+            label[0] for label in _get_pr_test_labels_from_event_payload(gh_event)
+        )
+
+    skip_code_coverage = True
+    if "test:coverage" in labels:
+        skip_code_coverage = False
+    elif event_name != "pull_request":
+        skip_code_coverage = False
+
     if event_name != "pull_request":
         # In this case, a full test run is in order
         ctx.info("Writing 'testrun' to the github outputs file")
-        testrun = {"type": "full"}
+        testrun = TestRun(type="full", skip_code_coverage=skip_code_coverage)
         with open(github_output, "a", encoding="utf-8") as wfh:
             wfh.write(f"testrun={json.dumps(testrun)}\n")
 
@@ -440,7 +499,7 @@ def define_testrun(ctx: Context, event_name: str, changed_files: pathlib.Path):
         ctx.exit(1)
 
     # So, it's a pull request...
-    # Based on which files changed, or other things like PR comments we can
+    # Based on which files changed, or other things like PR labels we can
     # decide what to run, or even if the full test run should be running on the
     # pull request, etc...
     changed_pkg_requirements_files = json.loads(
@@ -455,7 +514,7 @@ def define_testrun(ctx: Context, event_name: str, changed_files: pathlib.Path):
                 "Full test run chosen because there was a change made "
                 "to `cicd/golden-images.json`.\n"
             )
-        testrun = {"type": "full"}
+        testrun = TestRun(type="full", skip_code_coverage=skip_code_coverage)
     elif changed_pkg_requirements_files or changed_test_requirements_files:
         with open(github_step_summary, "a", encoding="utf-8") as wfh:
             wfh.write(
@@ -470,15 +529,20 @@ def define_testrun(ctx: Context, event_name: str, changed_files: pathlib.Path):
             ):
                 wfh.write(f"{path}\n")
             wfh.write("</pre>\n</details>\n")
-        testrun = {"type": "full"}
+        testrun = TestRun(type="full", skip_code_coverage=skip_code_coverage)
+    elif "test:full" in labels:
+        with open(github_step_summary, "a", encoding="utf-8") as wfh:
+            wfh.write("Full test run chosen because the label `test:full` is set.\n")
+        testrun = TestRun(type="full", skip_code_coverage=skip_code_coverage)
     else:
         testrun_changed_files_path = tools.utils.REPO_ROOT / "testrun-changed-files.txt"
-        testrun = {
-            "type": "changed",
-            "from-filenames": str(
+        testrun = TestRun(
+            type="changed",
+            skip_code_coverage=skip_code_coverage,
+            from_filenames=str(
                 testrun_changed_files_path.relative_to(tools.utils.REPO_ROOT)
             ),
-        }
+        )
         ctx.info(f"Writing {testrun_changed_files_path.name} ...")
         selected_changed_files = []
         for fpath in json.loads(changed_files_contents["testrun_files"]):
@@ -498,6 +562,28 @@ def define_testrun(ctx: Context, event_name: str, changed_files: pathlib.Path):
         if testrun["type"] == "changed":
             with open(github_step_summary, "a", encoding="utf-8") as wfh:
                 wfh.write("Partial test run chosen.\n")
+            testrun["selected_tests"] = {
+                "core": False,
+                "slow": False,
+                "fast": True,
+                "flaky": False,
+            }
+            if "test:slow" in labels:
+                with open(github_step_summary, "a", encoding="utf-8") as wfh:
+                    wfh.write("Slow tests chosen by `test:slow` label.\n")
+                testrun["selected_tests"]["slow"] = True
+            if "test:core" in labels:
+                with open(github_step_summary, "a", encoding="utf-8") as wfh:
+                    wfh.write("Core tests chosen by `test:core` label.\n")
+                testrun["selected_tests"]["core"] = True
+            if "test:no-fast" in labels:
+                with open(github_step_summary, "a", encoding="utf-8") as wfh:
+                    wfh.write("Fast tests deselected by `test:no-fast` label.\n")
+                testrun["selected_tests"]["fast"] = False
+            if "test:flaky-jail" in labels:
+                with open(github_step_summary, "a", encoding="utf-8") as wfh:
+                    wfh.write("Flaky jailed tests chosen by `test:flaky-jail` label.\n")
+                testrun["selected_tests"]["flaky"] = True
         if selected_changed_files:
             with open(github_step_summary, "a", encoding="utf-8") as wfh:
                 wfh.write(
@@ -585,18 +671,29 @@ def transport_matrix(ctx: Context, distro_slug: str):
             "help": "The distribution slug to generate the matrix for",
         },
         "pkg_type": {
-            "help": "The distribution slug to generate the matrix for",
+            "help": "The type of package we are testing against",
+        },
+        "testing_releases": {
+            "help": "The salt releases to test upgrades against",
+            "nargs": "+",
+            "required": True,
         },
     },
 )
-def pkg_matrix(ctx: Context, distro_slug: str, pkg_type: str):
+def pkg_matrix(
+    ctx: Context,
+    distro_slug: str,
+    pkg_type: str,
+    testing_releases: list[tools.utils.Version] = None,
+):
     """
     Generate the test matrix.
     """
     github_output = os.environ.get("GITHUB_OUTPUT")
     if github_output is None:
         ctx.warn("The 'GITHUB_OUTPUT' variable is not set.")
-
+    if TYPE_CHECKING:
+        assert testing_releases
     matrix = []
     sessions = [
         "install",
@@ -613,11 +710,12 @@ def pkg_matrix(ctx: Context, distro_slug: str, pkg_type: str):
         and pkg_type != "MSI"
     ):
         # These OS's never had arm64 packages built for them
-        # with the tiamate onedir packages.
+        # with the tiamat onedir packages.
         # we will need to ensure when we release 3006.0
         # we allow for 3006.0 jobs to run, because then
         # we will have arm64 onedir packages to upgrade from
         sessions.append("upgrade")
+    # TODO: Remove this block when we reach version 3009.0, we will no longer be testing upgrades from classic packages
     if (
         distro_slug
         not in [
@@ -633,11 +731,22 @@ def pkg_matrix(ctx: Context, distro_slug: str, pkg_type: str):
         sessions.append("upgrade-classic")
 
     for session in sessions:
-        matrix.append(
-            {
-                "test-chunk": session,
-            }
-        )
+        versions: list[str | None] = [None]
+        if session == "upgrade":
+            versions = [str(version) for version in testing_releases]
+        elif session == "upgrade-classic":
+            versions = [
+                str(version)
+                for version in testing_releases
+                if version < tools.utils.Version("3006.0")
+            ]
+        for version in versions:
+            matrix.append(
+                {
+                    "test-chunk": session,
+                    "version": version,
+                }
+            )
     ctx.info("Generated matrix:")
     ctx.print(matrix, soft_wrap=True)
 
@@ -672,3 +781,245 @@ def get_releases(ctx: Context, repository: str = "saltstack/salt"):
             wfh.write(f"latest-release={latest}\n")
             wfh.write(f"releases={json.dumps(str_releases)}\n")
         ctx.exit(0)
+
+
+@ci.command(
+    name="get-pr-test-labels",
+    arguments={
+        "pr": {
+            "help": "Pull request number",
+        },
+        "repository": {
+            "help": "Github repository.",
+        },
+    },
+)
+def get_pr_test_labels(
+    ctx: Context, repository: str = "saltstack/salt", pr: int = None
+):
+    """
+    Set the pull-request labels.
+    """
+    gh_event_path = os.environ.get("GITHUB_EVENT_PATH") or None
+    if gh_event_path is None:
+        labels = _get_pr_test_labels_from_api(ctx, repository, pr=pr)
+    else:
+        if TYPE_CHECKING:
+            assert gh_event_path is not None
+
+        try:
+            gh_event = json.loads(open(gh_event_path).read())
+        except Exception as exc:
+            ctx.error(
+                f"Could not load the GH Event payload from {gh_event_path!r}:\n", exc
+            )
+            ctx.exit(1)
+
+        if "pull_request" not in gh_event:
+            ctx.warning("The 'pull_request' key was not found on the event payload.")
+            ctx.exit(1)
+
+        pr = gh_event["pull_request"]["number"]
+        labels = _get_pr_test_labels_from_event_payload(gh_event)
+
+    if labels:
+        ctx.info(f"Test labels for pull-request #{pr} on {repository}:")
+        for name, description in labels:
+            ctx.info(f" * [yellow]{name}[/yellow]: {description}")
+    else:
+        ctx.info(f"No test labels for pull-request #{pr} on {repository}")
+
+    github_output = os.environ.get("GITHUB_OUTPUT")
+    if github_output is None:
+        ctx.exit(0)
+
+    if TYPE_CHECKING:
+        assert github_output is not None
+
+    ctx.info("Writing 'labels' to the github outputs file")
+    with open(github_output, "a", encoding="utf-8") as wfh:
+        wfh.write(f"labels={json.dumps([label[0] for label in labels])}\n")
+    ctx.exit(0)
+
+
+def _get_pr_test_labels_from_api(
+    ctx: Context, repository: str = "saltstack/salt", pr: int = None
+) -> list[tuple[str, str]]:
+    """
+    Set the pull-request labels.
+    """
+    if pr is None:
+        ctx.error(
+            "Could not find the 'GITHUB_EVENT_PATH' variable and the "
+            "--pr flag was not passed. Unable to detect pull-request number."
+        )
+        ctx.exit(1)
+    with ctx.web as web:
+        headers = {
+            "Accept": "application/vnd.github+json",
+        }
+        if "GITHUB_TOKEN" in os.environ:
+            headers["Authorization"] = f"Bearer {os.environ['GITHUB_TOKEN']}"
+        web.headers.update(headers)
+        ret = web.get(f"https://api.github.com/repos/{repository}/pulls/{pr}")
+        if ret.status_code != 200:
+            ctx.error(
+                f"Failed to get the #{pr} pull-request details on repository {repository!r}: {ret.reason}"
+            )
+            ctx.exit(1)
+        pr_details = ret.json()
+        return _filter_test_labels(pr_details["labels"])
+
+
+def _get_pr_test_labels_from_event_payload(
+    gh_event: dict[str, Any]
+) -> list[tuple[str, str]]:
+    """
+    Get the pull-request test labels.
+    """
+    if "pull_request" not in gh_event:
+        return []
+    return _filter_test_labels(gh_event["pull_request"]["labels"])
+
+
+def _filter_test_labels(labels: list[dict[str, Any]]) -> list[tuple[str, str]]:
+    return [
+        (label["name"], label["description"])
+        for label in labels
+        if label["name"].startswith("test:")
+    ]
+
+
+@ci.command(
+    name="get-testing-releases",
+    arguments={
+        "releases": {
+            "help": "The list of releases of salt",
+            "nargs": "*",
+        },
+        "salt_version": {
+            "help": "The version of salt being tested against",
+            "required": True,
+        },
+    },
+)
+def get_testing_releases(
+    ctx: Context,
+    releases: list[tools.utils.Version],
+    salt_version: str = None,
+):
+    """
+    Get a list of releases to use for the upgrade and downgrade tests.
+    """
+    github_output = os.environ.get("GITHUB_OUTPUT")
+    if github_output is None:
+        ctx.exit(1, "The 'GITHUB_OUTPUT' variable is not set.")
+    else:
+        # We aren't testing upgrades from anything before 3006.0 except the latest 3005.x
+        threshold_major = 3006
+        parsed_salt_version = tools.utils.Version(salt_version)
+        # We want the latest 4 major versions, removing the oldest if this version is a new major
+        num_major_versions = 4
+        if parsed_salt_version.minor == 0:
+            num_major_versions = 3
+        majors = sorted(
+            list(
+                {
+                    version.major
+                    for version in releases
+                    if version.major >= threshold_major
+                }
+            )
+        )[-num_major_versions:]
+        testing_releases = []
+        # Append the latest minor for each major
+        for major in majors:
+            minors_of_major = [
+                version for version in releases if version.major == major
+            ]
+            testing_releases.append(minors_of_major[-1])
+
+        # TODO: Remove this block when we reach version 3009.0
+        # Append the latest minor version of 3005 if we don't have enough major versions to test against
+        if len(testing_releases) != num_major_versions:
+            url = "https://repo.saltproject.io/salt/onedir/repo.json"
+            ret = ctx.web.get(url)
+            repo_data = ret.json()
+            latest = list(repo_data["latest"].keys())[0]
+            version = repo_data["latest"][latest]["version"]
+            testing_releases = [version] + testing_releases
+
+        str_releases = [str(version) for version in testing_releases]
+
+        with open(github_output, "a", encoding="utf-8") as wfh:
+            wfh.write(f"testing-releases={json.dumps(str_releases)}\n")
+
+        ctx.exit(0)
+
+
+@ci.command(
+    name="define-cache-seed",
+    arguments={
+        "static_cache_seed": {
+            "help": "The static cache seed value",
+        },
+        "randomize": {
+            "help": "Randomize the cache seed value",
+        },
+    },
+)
+def define_cache_seed(ctx: Context, static_cache_seed: str, randomize: bool = False):
+    """
+    Set `cache-seed` in GH Actions outputs.
+    """
+    github_output = os.environ.get("GITHUB_OUTPUT")
+    if github_output is None:
+        ctx.warn("The 'GITHUB_OUTPUT' variable is not set.")
+        ctx.exit(1)
+
+    if TYPE_CHECKING:
+        assert github_output is not None
+
+    github_step_summary = os.environ.get("GITHUB_STEP_SUMMARY")
+    if github_step_summary is None:
+        ctx.warn("The 'GITHUB_STEP_SUMMARY' variable is not set.")
+        ctx.exit(1)
+
+    if TYPE_CHECKING:
+        assert github_step_summary is not None
+
+    labels: list[str] = []
+    gh_event_path = os.environ.get("GITHUB_EVENT_PATH") or None
+    if gh_event_path is not None:
+        try:
+            gh_event = json.loads(open(gh_event_path).read())
+        except Exception as exc:
+            ctx.error(
+                f"Could not load the GH Event payload from {gh_event_path!r}:\n", exc
+            )
+            ctx.exit(1)
+
+        labels.extend(
+            label[0] for label in _get_pr_test_labels_from_event_payload(gh_event)
+        )
+
+    if randomize is True:
+        cache_seed = f"SEED-{random.randint(100, 1000)}"
+        with open(github_step_summary, "a", encoding="utf-8") as wfh:
+            wfh.write(
+                f"The cache seed has been randomized to `{cache_seed}` because "
+                "`--randomize` was passed to `tools ci define-cache-seed`."
+            )
+    elif "test:random-cache-seed" in labels:
+        cache_seed = f"SEED-{random.randint(100, 1000)}"
+        with open(github_step_summary, "a", encoding="utf-8") as wfh:
+            wfh.write(
+                f"The cache seed has been randomized to `{cache_seed}` because "
+                "the label `test:random-cache-seed` was set."
+            )
+    else:
+        cache_seed = static_cache_seed
+
+    ctx.info("Writing 'cache-seed' to the github outputs file")
+    with open(github_output, "a", encoding="utf-8") as wfh:
+        wfh.write(f"cache-seed={cache_seed}\n")
