+++ tools/pkg/build.py
@@ -80,11 +80,18 @@ def debian(
             ctx.exit(1)
         ctx.info("Building the package from the source files")
         shared_constants = _get_shared_constants()
+        if not python_version:
+            python_version = shared_constants["python_version"]
+        if not relenv_version:
+            relenv_version = shared_constants["relenv_version"]
+        if TYPE_CHECKING:
+            assert python_version
+            assert relenv_version
         new_env = {
-            "SALT_RELENV_VERSION": relenv_version or shared_constants["relenv_version"],
-            "SALT_PYTHON_VERSION": python_version
-            or shared_constants["python_version_linux"],
+            "SALT_RELENV_VERSION": relenv_version,
+            "SALT_PYTHON_VERSION": python_version,
             "SALT_PACKAGE_ARCH": str(arch),
+            "RELENV_FETCH_VERSION": relenv_version,
         }
         for key, value in new_env.items():
             os.environ[key] = value
@@ -140,11 +147,18 @@ def rpm(
             ctx.exit(1)
         ctx.info(f"Building the package from the source files")
         shared_constants = _get_shared_constants()
+        if not python_version:
+            python_version = shared_constants["python_version"]
+        if not relenv_version:
+            relenv_version = shared_constants["relenv_version"]
+        if TYPE_CHECKING:
+            assert python_version
+            assert relenv_version
         new_env = {
-            "SALT_RELENV_VERSION": relenv_version or shared_constants["relenv_version"],
-            "SALT_PYTHON_VERSION": python_version
-            or shared_constants["python_version_linux"],
+            "SALT_RELENV_VERSION": relenv_version,
+            "SALT_PYTHON_VERSION": python_version,
             "SALT_PACKAGE_ARCH": str(arch),
+            "RELENV_FETCH_VERSION": relenv_version,
         }
         for key, value in new_env.items():
             os.environ[key] = value
@@ -160,7 +174,6 @@ def rpm(
     arguments={
         "onedir": {
             "help": "The name of the onedir artifact, if given it should be under artifacts/",
-            "required": True,
         },
         "salt_version": {
             "help": (
@@ -172,10 +185,21 @@ def rpm(
         "sign": {
             "help": "Sign and notorize built package",
         },
+        "relenv_version": {
+            "help": "The version of relenv to use",
+        },
+        "python_version": {
+            "help": "The version of python to build with using relenv",
+        },
     },
 )
 def macos(
-    ctx: Context, onedir: str = None, salt_version: str = None, sign: bool = False
+    ctx: Context,
+    onedir: str = None,
+    salt_version: str = None,
+    sign: bool = False,
+    relenv_version: str = None,
+    python_version: str = None,
 ):
     """
     Build the macOS package.
@@ -185,15 +209,43 @@ def macos(
         assert salt_version is not None
 
     checkout = pathlib.Path.cwd()
-    onedir_artifact = checkout / "artifacts" / onedir
-    _check_pkg_build_files_exist(ctx, onedir_artifact=onedir_artifact)
+    if onedir:
+        onedir_artifact = checkout / "artifacts" / onedir
+        ctx.info(f"Building package from existing onedir: {str(onedir_artifact)}")
+        _check_pkg_build_files_exist(ctx, onedir_artifact=onedir_artifact)
 
-    build_root = checkout / "pkg" / "macos" / "build" / "opt"
-    build_root.mkdir(parents=True, exist_ok=True)
-    ctx.info(f"Extracting the onedir artifact to {build_root}")
-    with tarfile.open(str(onedir_artifact)) as tarball:
-        with ctx.chdir(onedir_artifact.parent):
-            tarball.extractall(path=build_root)
+        build_root = checkout / "pkg" / "macos" / "build" / "opt"
+        build_root.mkdir(parents=True, exist_ok=True)
+        ctx.info(f"Extracting the onedir artifact to {build_root}")
+        with tarfile.open(str(onedir_artifact)) as tarball:
+            with ctx.chdir(onedir_artifact.parent):
+                tarball.extractall(path=build_root)
+    else:
+        ctx.info("Building package without an existing onedir")
+
+    if not onedir:
+        # Prep the salt onedir if not building from an existing one
+        shared_constants = _get_shared_constants()
+        if not python_version:
+            python_version = shared_constants["python_version"]
+        if not relenv_version:
+            relenv_version = shared_constants["relenv_version"]
+        if TYPE_CHECKING:
+            assert python_version
+            assert relenv_version
+        os.environ["RELENV_FETCH_VERSION"] = relenv_version
+        with ctx.chdir(checkout / "pkg" / "macos"):
+            ctx.info("Fetching relenv python")
+            ctx.run(
+                "./build_python.sh",
+                "--version",
+                python_version,
+                "--relenv-version",
+                relenv_version,
+            )
+
+            ctx.info("Installing salt into the relenv python")
+            ctx.run("./install_salt.sh")
 
     if sign:
         ctx.info("Signing binaries")
@@ -222,7 +274,6 @@ def macos(
     arguments={
         "onedir": {
             "help": "The name of the onedir artifact, if given it should be under artifacts/",
-            "required": True,
         },
         "salt_version": {
             "help": (
@@ -237,7 +288,13 @@ def macos(
             "required": True,
         },
         "sign": {
-            "help": "Sign and notorize built package",
+            "help": "Sign and notarize built package",
+        },
+        "relenv_version": {
+            "help": "The version of relenv to use",
+        },
+        "python_version": {
+            "help": "The version of python to build with using relenv",
         },
     },
 )
@@ -247,33 +304,27 @@ def windows(
     salt_version: str = None,
     arch: str = None,
     sign: bool = False,
+    relenv_version: str = None,
+    python_version: str = None,
 ):
     """
     Build the Windows package.
     """
     if TYPE_CHECKING:
-        assert onedir is not None
         assert salt_version is not None
         assert arch is not None
 
-    checkout = pathlib.Path.cwd()
-    onedir_artifact = checkout / "artifacts" / onedir
-    _check_pkg_build_files_exist(ctx, onedir_artifact=onedir_artifact)
+    shared_constants = _get_shared_constants()
+    if not python_version:
+        python_version = shared_constants["python_version"]
+    if not relenv_version:
+        relenv_version = shared_constants["relenv_version"]
+    if TYPE_CHECKING:
+        assert python_version
+        assert relenv_version
+    os.environ["RELENV_FETCH_VERSION"] = relenv_version
 
-    unzip_dir = checkout / "pkg" / "windows"
-    ctx.info(f"Unzipping the onedir artifact to {unzip_dir}")
-    with zipfile.ZipFile(onedir_artifact, mode="r") as archive:
-        archive.extractall(unzip_dir)
-
-    move_dir = unzip_dir / "salt"
-    build_env = unzip_dir / "buildenv"
-    _check_pkg_build_files_exist(ctx, move_dir=move_dir)
-
-    ctx.info(f"Moving {move_dir} directory to the build environment in {build_env}")
-    shutil.move(move_dir, build_env)
-
-    ctx.info("Building the windows package")
-    ctx.run(
+    build_cmd = [
         "powershell.exe",
         "&",
         "pkg/windows/build.cmd",
@@ -281,9 +332,37 @@ def windows(
         arch,
         "-Version",
         salt_version,
+        "-PythonVersion",
+        python_version,
+        "-RelenvVersion",
+        relenv_version,
         "-CICD",
-        "-SkipInstall",
-    )
+    ]
+
+    checkout = pathlib.Path.cwd()
+    if onedir:
+        build_cmd.append("-SkipInstall")
+        onedir_artifact = checkout / "artifacts" / onedir
+        ctx.info(f"Building package from existing onedir: {str(onedir_artifact)}")
+        _check_pkg_build_files_exist(ctx, onedir_artifact=onedir_artifact)
+
+        unzip_dir = checkout / "pkg" / "windows"
+        ctx.info(f"Unzipping the onedir artifact to {unzip_dir}")
+        with zipfile.ZipFile(onedir_artifact, mode="r") as archive:
+            archive.extractall(unzip_dir)
+
+        move_dir = unzip_dir / "salt"
+        build_env = unzip_dir / "buildenv"
+        _check_pkg_build_files_exist(ctx, move_dir=move_dir)
+
+        ctx.info(f"Moving {move_dir} directory to the build environment in {build_env}")
+        shutil.move(move_dir, build_env)
+    else:
+        build_cmd.append("-Build")
+        ctx.info("Building package without an existing onedir")
+
+    ctx.info(f"Running: {' '.join(build_cmd)} ...")
+    ctx.run(*build_cmd)
 
     if sign:
         env = os.environ.copy()
@@ -364,6 +443,9 @@ def windows(
             "help": "The version of python to create an environment for using relenv",
             "required": True,
         },
+        "relenv_version": {
+            "help": "The version of relenv to use",
+        },
         "package_name": {
             "help": "The name of the relenv environment to be created",
             "required": True,
@@ -378,6 +460,7 @@ def onedir_dependencies(
     ctx: Context,
     arch: str = None,
     python_version: str = None,
+    relenv_version: str = None,
     package_name: str = None,
     platform: str = None,
 ):
@@ -391,6 +474,16 @@ def onedir_dependencies(
         assert python_version is not None
         assert package_name is not None
         assert platform is not None
+
+    shared_constants = _get_shared_constants()
+    if not python_version:
+        python_version = shared_constants["python_version"]
+    if not relenv_version:
+        relenv_version = shared_constants["relenv_version"]
+    if TYPE_CHECKING:
+        assert python_version
+        assert relenv_version
+    os.environ["RELENV_FETCH_VERSION"] = relenv_version
 
     # We import relenv here because it is not a hard requirement for the rest of the tools commands
     try:
@@ -414,12 +507,11 @@ def onedir_dependencies(
         ctx.error(f"Failed to get the relenv version: {ret}")
         ctx.exit(1)
 
-    target_relenv_version = _get_shared_constants()["relenv_version"]
     env_relenv_version = ret.stdout.strip().decode()
-    if env_relenv_version != target_relenv_version:
+    if env_relenv_version != relenv_version:
         ctx.error(
             f"The onedir installed relenv version({env_relenv_version}) is not "
-            f"the relenv version which should be used({target_relenv_version})."
+            f"the relenv version which should be used({relenv_version})."
         )
         ctx.exit(1)
 
@@ -510,6 +602,9 @@ def onedir_dependencies(
             "help": "The name of the relenv environment to install salt into",
             "required": True,
         },
+        "relenv_version": {
+            "help": "The version of relenv to use",
+        },
     },
 )
 def salt_onedir(
@@ -517,6 +612,7 @@ def salt_onedir(
     salt_name: str,
     platform: str = None,
     package_name: str = None,
+    relenv_version: str = None,
 ):
     """
     Install salt into a relenv onedir environment.
@@ -524,6 +620,13 @@ def salt_onedir(
     if TYPE_CHECKING:
         assert platform is not None
         assert package_name is not None
+
+    shared_constants = _get_shared_constants()
+    if not relenv_version:
+        relenv_version = shared_constants["relenv_version"]
+    if TYPE_CHECKING:
+        assert relenv_version
+    os.environ["RELENV_FETCH_VERSION"] = relenv_version
 
     salt_archive = pathlib.Path(salt_name).resolve()
     onedir_env = pathlib.Path(package_name).resolve()
@@ -542,12 +645,11 @@ def salt_onedir(
         ctx.error(f"Failed to get the relenv version: {ret}")
         ctx.exit(1)
 
-    target_relenv_version = _get_shared_constants()["relenv_version"]
     env_relenv_version = ret.stdout.strip().decode()
-    if env_relenv_version != target_relenv_version:
+    if env_relenv_version != relenv_version:
         ctx.error(
             f"The onedir installed relenv version({env_relenv_version}) is not "
-            f"the relenv version which should be used({target_relenv_version})."
+            f"the relenv version which should be used({relenv_version})."
         )
         ctx.exit(1)
